Start testing main.py
