+++ src/zope/app/server/tests/test_mkzopeinstance.py
@@ -16,31 +16,11 @@
 
 import os
 import shutil
-import sys
 import tempfile
 import unittest
-from contextlib import contextmanager
-from io import StringIO, BytesIO
 
 from zope.app.server import mkzopeinstance
-
-
-# By using io.BytesIO() instead of cStringIO.StringIO() on Python 2 we make
-# sure we're not trying to accidentally print unicode to stdout/stderr.
-NativeStringIO = BytesIO if str is bytes else StringIO
-
-
-@contextmanager
-def capture_output(stdout=None, stderr=None):
-    old_stdout = sys.stdout
-    old_stderr = sys.stderr
-    sys.stdout = stdout = stdout or NativeStringIO()
-    sys.stderr = stderr = stderr or NativeStringIO()
-    try:
-        yield stdout, stderr
-    finally:
-        sys.stdout = old_stdout
-        sys.stderr = old_stderr
+from zope.app.server._testing import capture_output
 
 
 class ArgumentParsingTestCase(unittest.TestCase):
